# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Grid Origination Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ercot-node-stats:
    get:
      operationId: listErcotNodeStats
      tags: [ercot-node-stats]
      summary: List ERCOT hub/zone monthly stats
      parameters:
        - name: node
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: ERCOT node stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErcotNodeStats"
  /ercot-settlement-points:
    get:
      operationId: listErcotSettlementPoints
      tags: [ercot-nodal-stats]
      summary: List distinct ERCOT settlement point names (resource nodes only)
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of settlement point names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /ercot-nodal-stats:
    get:
      operationId: listErcotNodalStats
      tags: [ercot-nodal-stats]
      summary: List ERCOT resource nodal stats
      parameters:
        - name: settlementPoint
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: ERCOT nodal stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErcotNodalStats"
  /caiso-node-stats:
    get:
      operationId: listCaisoNodeStats
      tags: [caiso-node-stats]
      summary: List CAISO node stats
      parameters:
        - name: node
          in: query
          schema:
            type: string
            enum: [NP15, SP15, ZP26]
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: CAISO node stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaisoNodeStats"
  /pjm-node-stats:
    get:
      operationId: listPjmNodeStats
      tags: [pjm-node-stats]
      summary: List PJM hub/zone node stats
      parameters:
        - name: node
          in: query
          schema:
            type: string
            enum: [Western Hub, Eastern Hub, AEP-Dayton Hub, NI Hub, PSEG, PPL, DOM, BGE]
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: PJM node stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PjmNodeStats"
...
components:
  schemas:
...
    ErcotNodeStats:
      type: object
      properties:
        id:
          type: integer
        node:
          type: string
        nodeType:
          type: string
          enum: [hub, load_zone]
        year:
          type: integer
        month:
          type: integer
        avgDaPrice:
          type: number
        avgRtPrice:
          type: number
        volatility:
          type: number
        negPricePercent:
          type: number
        onPeakAvg:
          type: number
        offPeakAvg:
          type: number
        minPrice:
          type: number
        maxPrice:
          type: number
      required: [id, node, nodeType, year, month, avgDaPrice, avgRtPrice]
    ErcotNodalStats:
      type: object
      properties:
        id:
          type: integer
        settlementPoint:
          type: string
        year:
          type: integer
        month:
          type: integer
        avgDaPrice:
          type: number
        avgRtPrice:
          type: number
        stdDev:
          type: number
        negPricePercent:
          type: number
        onPeakAvg:
          type: number
        offPeakAvg:
          type: number
        minPrice:
          type: number
        maxPrice:
          type: number
        sampleCount:
          type: integer
      required: [id, settlementPoint, year, month, avgDaPrice]
    CaisoNodeStats:
      type: object
      properties:
        id:
          type: integer
        node:
          type: string
          enum: [NP15, SP15, ZP26]
        year:
          type: integer
        month:
          type: integer
        avgDaPrice:
          type: number
        avgRtPrice:
          type: number
        volatility:
          type: number
        negPricePercent:
          type: number
        onPeakAvg:
          type: number
        offPeakAvg:
          type: number
      required: [id, node, year, month, avgDaPrice, avgRtPrice]
    PjmNodeStats:
      type: object
      properties:
        id:
          type: integer
        node:
          type: string
          enum: [Western Hub, Eastern Hub, AEP-Dayton Hub, NI Hub, PSEG, PPL, DOM, BGE]
        year:
          type: integer
        month:
          type: integer
        avgDaPrice:
          type: number
        avgRtPrice:
          type: number
        volatility:
          type: number
        negPricePercent:
          type: number
        onPeakAvg:
          type: number
        offPeakAvg:
          type: number
      required: [id, node, year, month, avgDaPrice]
...